# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shelfkeeper API
  version: 0.0.1
  description: |
    図書館向けの蔵書・利用者・貸出管理 API。

    管理者は `POST /login` で JWT を取得し、以降のリクエストに
    `Authorization: Bearer <token>` を付与する。`POST /login` 以外は認証必須。
  contact:
    name: itokanta
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: 利用者一覧を取得する
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
              example:
                userList:
                  - id: 1
                    name: 佐藤花子
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Users]
      operationId: createUser
      summary: 利用者を登録する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
            example:
              name: 佐藤花子
      responses:
        "201":
          description: 作成成功。レスポンスボディなし
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 400
                message: バリデーションエラー
                details:
                  name: 名前は必須です
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/overduelist:
    get:
      tags: [Users]
      operationId: listOverdueUsers
      summary: 延滞利用者一覧を取得する
      description: 返却期限を過ぎた未返却貸出がある利用者と、その延滞書籍を返す。
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOverDueListResponse"
              example:
                userOverDueList:
                  - userId: 1
                    userName: 佐藤花子
                    dueBookList:
                      - bookId: 2
                        bookTitle: 羅生門
                        loanDate: "2026-08-01"
                        dueDate: "2026-08-15"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{id}:
    put:
      tags: [Users]
      operationId: updateUser
      summary: 利用者を更新する
      description: 送信した項目だけを更新する。省略した項目は既存値を維持する。
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
            example:
              name: 佐藤花子
      responses:
        "204":
          description: 更新成功。レスポンスボディなし
        "400":
          description: |
            パスパラメータ不正、または指定利用者が未登録。
            未登録時の `message` は「指定されたユーザーは存在しません」。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validation:
                  summary: ID が 1 未満
                  value:
                    status: 400
                    message: バリデーションエラー
                    details:
                      id: ユーザーIDは1以上を指定してください
                notFound:
                  summary: 利用者が存在しない
                  value:
                    status: 400
                    message: 指定されたユーザーは存在しません
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Users]
      operationId: deleteUser
      summary: 利用者を削除する
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "204":
          description: 削除成功。レスポンスボディなし
        "400":
          description: |
            パスパラメータ不正、または指定利用者が未登録。
            未登録時の `message` は「指定されたユーザーは存在しません」。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validation:
                  summary: ID が 1 未満
                  value:
                    status: 400
                    message: バリデーションエラー
                    details:
                      id: ユーザーIDは1以上を指定してください
                notFound:
                  summary: 利用者が存在しない
                  value:
                    status: 400
                    message: 指定されたユーザーは存在しません
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "POST /login で取得した JWT を Authorization ヘッダーの Bearer トークンとして送る。"
  parameters:
...
    UserIdPath:
      name: id
      in: path
      required: true
      description: 利用者の識別子
      schema:
        type: integer
        format: int32
        minimum: 1
...
  responses:
    Unauthorized:
      description: |
        未認証、またはトークンが無効。
        トークン未送信時はボディなし。無効トークン時は `{"message":"トークンが無効です"}`。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthErrorResponse"
          example:
            message: トークンが無効です
    InternalServerError:
      description: 想定外のサーバーエラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 500
            message: サーバーエラーが発生しました
  schemas:
...
    AuthErrorResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
          format: int32
          description: HTTP ステータスコード
        message:
          type: string
          description: エラー概要
        details:
          type: object
          additionalProperties:
            type: string
          description: 項目ごとのバリデーションメッセージ。業務エラーでは省略される。
          nullable: true
...
    UserCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          description: 利用者名
    UserUpdateRequest:
      type: object
      description: 更新する項目だけを送る。未送信の項目は変更しない。
      properties:
        name:
          type: string
          description: 利用者名
    UserListResponse:
      type: object
      required: [userList]
      properties:
        userList:
          type: array
          items:
            $ref: "#/components/schemas/UserListItem"
    UserListItem:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int32
          description: 利用者の識別子
        name:
          type: string
          description: 利用者名
    UserOverDueListResponse:
      type: object
      required: [userOverDueList]
      properties:
        userOverDueList:
          type: array
          items:
            $ref: "#/components/schemas/UserOverDueListItem"
    UserOverDueListItem:
      type: object
      required: [userId, userName, dueBookList]
      properties:
        userId:
          type: integer
          format: int32
          description: 利用者の識別子
        userName:
          type: string
          description: 利用者名
        dueBookList:
          type: array
          description: 延滞中の書籍一覧
          items:
            $ref: "#/components/schemas/DueBookListItem"
    DueBookListItem:
      type: object
      required: [bookId, bookTitle, loanDate, dueDate]
      properties:
        bookId:
          type: integer
          format: int32
          description: 書籍の識別子
        bookTitle:
          type: string
          description: 書籍のタイトル
        loanDate:
          type: string
          format: date
          description: 貸出日
        dueDate:
          type: string
          format: date
          description: 返却期限日
...